Refactor: Move mappers to domain.mappers package
Moves mapper functions and related constants from the top-level `com.gabrielbmoro.moviedb.domain` package to `com.gabrielbmoro.moviedb.domain.mappers`.

This improves code organization and modularity within the domain layer.

Updated imports in affected files and adjusted Popcorn configuration for the new package structure.
